chatbot/requirements.txt
langchain
langchain-google-genai
langchain-core
langchain-community
langchain-tavily
langgraph
mcp
langchain-mcp-adapters
fastapi
uvicorn
python-dotenv
rich
pydantic
